refactor: Move viewmodels to domain layer.
